requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
appnope==0.1.2
asttokens==2.0.5
async-timeout==4.0.1
asynctest==0.13.0
attrs==21.2.0
backcall==0.2.0
base58==2.1.1
bitarray==1.2.2
black==21.11b1
blessings==1.7
bpython==0.22.1
cached-property==1.5.2
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
curtsies==0.3.10
cwcwidth==0.1.6
cytoolz==0.11.2
dataclassy==0.11.1
decorator==5.1.1
eip712==0.1.0
eth-abi==2.1.1
eth-account==0.5.6
eth-brownie==1.17.2
eth-event==1.2.3
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
execnet==1.9.0
executing==0.8.2
frozenlist==1.2.0
greenlet==1.1.2
hexbytes==0.2.2
hypothesis==6.27.3
idna==3.3
importlib-metadata==4.8.2
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
ipython==8.0.1
jedi==0.18.1
jsonschema==3.2.0
lazy-object-proxy==1.6.0
lru-dict==1.1.7
matplotlib-inline==0.1.3
multiaddr==0.0.9
multidict==5.2.0
mypy-extensions==0.4.3
mythx-models==1.9.1
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.0
pluggy==1.0.0
prompt-toolkit==3.0.23
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pycryptodome==3.12.0
Pygments==2.10.0
pygments-lexer-solidity==0.7.0
PyJWT==1.7.1
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==1.34.0
python-dateutil==2.8.1
python-dotenv==0.16.0
pythx==1.6.1
pyxdg==0.27
PyYAML==5.4.1
regex==2021.11.10
requests==2.26.0
rlp==2.0.1
semantic-version==2.8.5
six==1.16.0
sortedcontainers==2.4.0
stack-data==0.1.4
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tqdm==4.62.3
traitlets==5.1.1
typed-ast==1.5.1
typing-extensions==3.10.0.2
urllib3==1.26.7
varint==1.0.2
vvm==0.1.0
vyper==0.3.1
wcwidth==0.2.5
web3==5.25.0
websockets==9.1
wrapt==1.13.3
yarl==1.7.2
zipp==3.6.0
